stage1/requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
trl>=0.8.0
peft>=0.10.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
pyyaml
jsonlines>=4.0.0
